Update README.md
The Resume Builder Project using Bootstrap 5 is an innovative web application that empowers users to create polished and professional resumes effortlessly. Built on the robust foundation of Bootstrap 5, this project offers an array of modern resume templates and a rich set of features to elevate the resume-building experience. Whether you're a job seeker striving to craft the perfect resume or a developer eager to explore the capabilities of Bootstrap 5, this project serves as a valuable resource.